修改 类型转换的错误

+++ Brochure.Core/Abstracts/EntityBase.cs
@@ -8,13 +8,13 @@ namespace Brochure.Core
         /// 主建
         /// </summary>
         /// <returns></returns>
-        public Guid Id { get; set; } = Guid.NewGuid ();
+        public Guid Id { get; set; } = Guid.NewGuid();
         /// <summary>
         /// 创建时间
         /// </summary>
         /// <returns></returns>
         public DateTime CreateTime { get; set; } = DateTime.Now;
         //序列号
-        public int SequenceId { get; set; }
+        public long SequenceId { get; set; }
     }
 }
